split out tests of layered dfsu to new file
