+++ tests/test_dfsu.py
@@ -7,7 +7,6 @@ import pandas as pd
 import pytest
 import mikeio
 from mikeio import Dataset, DataArray, Dfs0, Dfsu, Mesh
-from mikeio.custom_exceptions import InvalidGeometry
 from mikeio.eum import ItemInfo
 from pytest import approx
 
@@ -86,32 +85,6 @@ def test_read_timestep_1():
     assert len(ds.time) == 1
 
 
-def test_read_simple_3d():
-    filename = os.path.join("tests", "testdata", "basin_3d.dfsu")
-    dfs = Dfsu(filename)
-
-    ds = dfs.read()
-
-    assert len(ds.data) == 3
-    assert len(ds.items) == 3
-
-    assert ds.items[0].name != "Z coordinate"
-    assert ds.items[2].name == "W velocity"
-
-
-def test_read_simple_2dv():
-    filename = os.path.join("tests", "testdata", "basin_2dv.dfsu")
-    dfs = Dfsu(filename)
-
-    ds = dfs.read()
-
-    assert len(ds.data) == 3
-    assert len(ds.items) == 3
-
-    assert ds.items[0].name != "Z coordinate"
-    assert ds.items[2].name == "W velocity"
-
-
 def test_read_single_item_returns_single_item():
     filename = "tests/testdata/HD2D.dfsu"
     dfs = Dfsu(filename)
@@ -159,18 +132,6 @@ def test_read_selected_item_names_returns_correct_items():
     assert len(ds) == 2
     assert ds.items[0].name == "Surface elevation"
     assert ds.items[1].name == "Current speed"
-
-
-def test_read_returns_correct_items_sigma_z():
-    filename = "tests/testdata/oresund_sigma_z.dfsu"
-    dfs = Dfsu(filename)
-
-    ds = dfs.read()
-
-    assert len(ds) == 2
-    # assert ds.items[0].name == "Z coordinate"
-    assert ds.items[0].name == "Temperature"
-    assert ds.items[1].name == "Salinity"
 
 
 def test_read_all_time_steps():
@@ -249,14 +210,6 @@ def test_number_of_time_steps():
     assert dfs.n_timesteps == 9
 
 
-def test_number_of_nodes_and_elements_sigma_z():
-    filename = "tests/testdata/oresund_sigma_z.dfsu"
-    dfs = Dfsu(filename)
-
-    assert dfs.n_elements == 17118
-    assert dfs.n_nodes == 12042
-
-
 def test_get_node_coords():
     filename = "tests/testdata/HD2D.dfsu"
     dfs = Dfsu(filename)
@@ -274,19 +227,6 @@ def test_element_coordinates():
 
     ec = dfs.element_coordinates
     assert ec[1, 1] == pytest.approx(6906790.5928664245)
-
-
-def test_calc_element_coordinates_3d():
-    filename = "tests/testdata/oresund_sigma_z.dfsu"
-    dfs = Dfsu(filename)
-
-    # extract dynamic z values for profile
-    elem_ids = dfs.find_nearest_profile_elements(333934.1, 6158101.5)
-    ds = dfs.read(items=0, elements=elem_ids, time=0)
-    zn_dyn = ds[0]._zn  # TODO
-    ec = dfs.calc_element_coordinates(elements=elem_ids, zn=zn_dyn)
-
-    assert ec[0, 2] == pytest.approx(-6.981768845)
 
 
 def test_element_coords_is_inside_nodes():
@@ -407,21 +347,6 @@ def test_find_nearest_elements_2d_array():
     assert elem_ids[1] == 317
 
 
-def test_find_nearest_elements_3d():
-    filename = os.path.join("tests", "testdata", "oresund_sigma_z.dfsu")
-    dfs = Dfsu(filename)
-
-    elem_id = dfs.find_nearest_elements(333934, 6158101)
-    assert elem_id == 5323
-    assert elem_id in dfs.top_elements
-
-    elem_id = dfs.find_nearest_elements(333934, 6158101, layer=7)
-    assert elem_id == 5322
-
-    elem_id = dfs.find_nearest_elements(333934, 6158101, -7)
-    assert elem_id == 5320
-
-
 def find_nearest_profile_elements():
     filename = os.path.join("tests", "testdata", "oresund_sigma_z.dfsu")
     dfs = Dfsu(filename)
@@ -447,18 +372,6 @@ def test_read_and_select_single_element():
     assert selds.data[0].shape == (9,)
 
 
-def test_read_and_select_single_element_dfsu_3d():
-
-    filename = os.path.join("tests", "testdata", "basin_3d.dfsu")
-    dfs = Dfsu(filename)
-
-    ds = dfs.read()
-
-    selds = ds.isel(idx=1739, axis=1)
-
-    assert selds.data[0].shape == (3,)
-
-
 def test_is_2d():
 
     filename = "tests/testdata/HD2D.dfsu"
@@ -470,168 +383,6 @@ def test_is_2d():
     dfs = Dfsu(filename)
 
     assert not dfs.is_2d
-
-
-def test_n_layers():
-
-    filename = os.path.join("tests", "testdata", "basin_3d.dfsu")
-    dfs = Dfsu(filename)
-    assert dfs.n_layers == 10
-
-    filename = os.path.join("tests", "testdata", "oresund_sigma_z.dfsu")
-    dfs = Dfsu(filename)
-    assert dfs.n_layers == 9
-
-    filename = os.path.join("tests", "testdata", "oresund_vertical_slice.dfsu")
-    dfs = Dfsu(filename)
-    assert dfs.n_layers == 9
-
-    filename = "tests/testdata/HD2D.dfsu"
-    dfs = Dfsu(filename)
-    assert not hasattr(dfs, "n_layers")
-
-
-def test_n_sigma_layers():
-
-    filename = os.path.join("tests", "testdata", "basin_3d.dfsu")
-    dfs = Dfsu(filename)
-    assert dfs.n_sigma_layers == 10
-
-    filename = os.path.join("tests", "testdata", "oresund_sigma_z.dfsu")
-    dfs = Dfsu(filename)
-    assert dfs.n_sigma_layers == 4
-
-    filename = os.path.join("tests", "testdata", "oresund_vertical_slice.dfsu")
-    dfs = Dfsu(filename)
-    assert dfs.n_sigma_layers == 4
-
-    filename = "tests/testdata/HD2D.dfsu"
-    dfs = Dfsu(filename)
-    assert not hasattr(dfs, "n_sigma_layers")
-
-
-def test_n_z_layers():
-
-    filename = os.path.join("tests", "testdata", "basin_3d.dfsu")
-    dfs = Dfsu(filename)
-    assert dfs.n_z_layers == 0
-
-    filename = os.path.join("tests", "testdata", "oresund_sigma_z.dfsu")
-    dfs = Dfsu(filename)
-    assert dfs.n_z_layers == 5
-
-    filename = os.path.join("tests", "testdata", "oresund_vertical_slice.dfsu")
-    dfs = Dfsu(filename)
-    assert dfs.n_z_layers == 5
-
-    filename = "tests/testdata/HD2D.dfsu"
-    dfs = Dfsu(filename)
-    assert not hasattr(dfs, "n_z_layers")
-
-
-def test_boundary_codes():
-
-    filename = os.path.join("tests", "testdata", "basin_3d.dfsu")
-    dfs = Dfsu(filename)
-    assert len(dfs.boundary_codes) == 1
-
-    filename = os.path.join("tests", "testdata", "oresund_sigma_z.dfsu")
-    dfs = Dfsu(filename)
-
-    assert len(dfs.boundary_codes) == 3
-
-
-def test_top_elements():
-    filename = os.path.join("tests", "testdata", "basin_3d.dfsu")
-    dfs = Dfsu(filename)
-    assert len(dfs.top_elements) == 174
-    assert dfs.top_elements[3] == 39
-
-    filename = os.path.join("tests", "testdata", "oresund_sigma_z.dfsu")
-    dfs = Dfsu(filename)
-    assert len(dfs.top_elements) == 3700
-    assert dfs.top_elements[3] == 16
-
-    filename = os.path.join("tests", "testdata", "oresund_vertical_slice.dfsu")
-    dfs = Dfsu(filename)
-    assert len(dfs.top_elements) == 99
-    assert dfs.top_elements[3] == 19
-
-    filename = "tests/testdata/HD2D.dfsu"
-    dfs = Dfsu(filename)
-    assert not hasattr(dfs, "top_elements")
-
-
-def test_bottom_elements():
-    filename = os.path.join("tests", "testdata", "basin_3d.dfsu")
-    dfs = Dfsu(filename)
-    assert len(dfs.bottom_elements) == 174
-    assert dfs.bottom_elements[3] == 30
-
-    filename = os.path.join("tests", "testdata", "oresund_sigma_z.dfsu")
-    dfs = Dfsu(filename)
-    assert len(dfs.bottom_elements) == 3700
-    assert dfs.bottom_elements[3] == 13
-
-    filename = os.path.join("tests", "testdata", "oresund_vertical_slice.dfsu")
-    dfs = Dfsu(filename)
-    assert len(dfs.bottom_elements) == 99
-    assert dfs.bottom_elements[3] == 15
-
-    filename = "tests/testdata/HD2D.dfsu"
-    dfs = Dfsu(filename)
-    assert not hasattr(dfs, "bottom_elements")
-
-
-def test_n_layers_per_column():
-    filename = os.path.join("tests", "testdata", "basin_3d.dfsu")
-    dfs = Dfsu(filename)
-    assert len(dfs.n_layers_per_column) == 174
-    assert dfs.n_layers_per_column[3] == 10
-
-    filename = os.path.join("tests", "testdata", "oresund_sigma_z.dfsu")
-    dfs = Dfsu(filename)
-    assert len(dfs.n_layers_per_column) == 3700
-    assert dfs.n_layers_per_column[3] == 4
-    assert max(dfs.n_layers_per_column) == dfs.n_layers
-
-    filename = os.path.join("tests", "testdata", "oresund_vertical_slice.dfsu")
-    dfs = Dfsu(filename)
-    assert len(dfs.n_layers_per_column) == 99
-    assert dfs.n_layers_per_column[3] == 5
-
-    filename = "tests/testdata/HD2D.dfsu"
-    dfs = Dfsu(filename)
-    assert not hasattr(dfs, "n_layers_per_column")
-
-
-def test_get_layer_elements():
-    filename = os.path.join("tests", "testdata", "oresund_sigma_z.dfsu")
-    dfs = Dfsu(filename)
-
-    elem_ids = dfs.get_layer_elements(-1)
-    assert np.all(elem_ids == dfs.top_elements)
-
-    elem_ids = dfs.get_layer_elements(-2)
-    assert elem_ids[5] == 23
-
-    elem_ids = dfs.get_layer_elements(0)
-    assert elem_ids[5] == 8638
-    assert len(elem_ids) == 10
-
-    elem_ids = dfs.get_layer_elements([0, 2])
-    assert len(elem_ids) == 197
-
-    with pytest.raises(Exception):
-        elem_ids = dfs.get_layer_elements(11)
-
-
-def test_find_nearest_profile_elements():
-    filename = os.path.join("tests", "testdata", "oresund_sigma_z.dfsu")
-    dfs = Dfsu(filename)
-    elem_ids = dfs.find_nearest_profile_elements(358337, 6196090)
-    assert len(elem_ids) == 8
-    assert elem_ids[-1] == 3042
 
 
 def test_is_geo_UTM():
@@ -657,13 +408,6 @@ def test_get_element_area_UTM():
     dfs = Dfsu(filename)
     areas = dfs.get_element_area()
     assert areas[0] == 4949.102548750438
-
-
-def test_get_element_area_3D():
-    filename = os.path.join("tests", "testdata", "oresund_sigma_z.dfsu")
-    dfs = Dfsu(filename)
-    areas = dfs.get_element_area()
-    assert areas[0] == 350186.43530453625
 
 
 def test_get_element_area_LONGLAT():
@@ -847,33 +591,6 @@ def test_write_from_dfsu_default_items_numbered(tmpdir):
     assert newdfs.items[0].name == "Item 1"
 
 
-def test_write_from_dfsu3D(tmpdir):
-
-    sourcefilename = os.path.join("tests", "testdata", "basin_3d.dfsu")
-    outfilename = os.path.join(tmpdir.dirname, "simple3D.dfsu")
-    dfs = Dfsu(sourcefilename)
-
-    ds = dfs.read(items=[0, 1])
-
-    dfs.write(outfilename, ds)
-
-    assert os.path.exists(outfilename)
-
-
-# def test_write_from_dfsu3D_withou_z_coords_fails(tmpdir):
-
-#     sourcefilename = os.path.join("tests", "testdata", "basin_3d.dfsu")
-#     outfilename = os.path.join(tmpdir.dirname, "simple3D_fail.dfsu")
-#     dfs = Dfsu(sourcefilename)
-
-#     ds = dfs.read([1, 2])
-#     with pytest.raises(Exception) as excinfo:
-#         dfs.write(outfilename, ds)
-
-#     assert "coord" in str(excinfo.value)
-#     assert not os.path.exists(outfilename)
-
-
 def test_write_invalid_data_closes_and_deletes_file(tmpdir):
 
     filename = os.path.join(tmpdir.dirname, "simple.dfsu")
@@ -995,23 +712,6 @@ def test_write_temporal_subset(tmpdir):
     assert newdfs.n_timesteps == 7
 
 
-def test_extract_top_layer_to_2d(tmpdir):
-    filename = os.path.join("tests", "testdata", "oresund_sigma_z.dfsu")
-
-    dfs = Dfsu(filename)
-    top_ids = dfs.top_elements
-
-    ds = dfs.read(elements=top_ids)
-
-    outfilename = os.path.join(tmpdir, "toplayer.dfsu")
-    dfs.write(outfilename, ds, elements=top_ids)
-
-    newdfs = Dfsu(outfilename)
-    assert os.path.exists(outfilename)
-
-    assert newdfs.is_2d
-
-
 def test_geometry_2d():
 
     filename = os.path.join("tests", "testdata", "oresund_sigma_z.dfsu")
@@ -1049,23 +749,6 @@ def test_geometry_2d():
 
 #     with pytest.raises(InvalidGeometry):
 #         dfs.find_nearest_profile_elements(x=0, y=0)
-
-
-def test_to_mesh_3d(tmpdir):
-
-    filename = os.path.join("tests", "testdata", "oresund_sigma_z.dfsu")
-
-    dfs = Dfsu(filename)
-
-    outfilename = os.path.join(tmpdir, "oresund.mesh")
-
-    dfs.to_mesh(outfilename)
-
-    assert os.path.exists(outfilename)
-
-    mesh = Mesh(outfilename)
-
-    assert True
 
 
 def test_to_mesh_2d(tmpdir):
@@ -1204,41 +887,6 @@ def test_extract_bad_track():
         dfs.extract_track(df)
 
 
-def test_extract_surface_elevation_from_3d():
-    dfs = Dfsu("tests/testdata/oresund_sigma_z.dfsu")
-    outputfile = "tests/testdata/oresund_surface_elev_extracted.dfsu"
-    n_top1 = len(dfs.top_elements)
-
-    dfs.extract_surface_elevation_from_3d(outputfile, time=-1)
-
-    dfs2 = Dfsu(outputfile)
-    assert dfs2.n_elements == n_top1
-    os.remove(outputfile)  # clean up
-
-
-def test_find_nearest_element_in_Zlayer():
-    filename = os.path.join("tests", "testdata", "oresund_sigma_z.dfsu")
-    dfs = Dfsu(filename)
-    el2dindx = dfs.elem2d_ids[12]
-    assert el2dindx == 2
-    ids = dfs.find_nearest_elements(357000, 6200000, layer=0)
-    el2dindx = dfs.elem2d_ids[ids]
-    table = dfs.e2_e3_table[el2dindx]
-    assert ids == 3216
-    assert el2dindx == 745
-    assert len(table) == 9
-    ids = dfs.find_nearest_elements(357000, 6200000, layer=8)
-    el2dindx = dfs.elem2d_ids[ids]
-    table = dfs.e2_e3_table[el2dindx]
-    assert ids == 3224
-    assert el2dindx == 745
-    assert len(table) == 9
-
-    with pytest.raises(Exception):
-        # z and layer cannot both be given
-        dfs.find_nearest_elements(357000, 6200000, z=-3, layer=8)
-
-
 def test_e2_e3_table_2d_file():
     filename = os.path.join("tests", "testdata", "NorthSea_HD_and_windspeed.dfsu")
     dfs = Dfsu(filename)
@@ -1296,30 +944,6 @@ def test_dataset_write_dfsu(tmp_path):
 
     ds2 = mikeio.read(outfilename)
     assert ds2.n_timesteps == 2
-
-
-def test_dataset_write_dfsu3d(tmp_path):
-
-    outfilename = tmp_path / "oresund_sigma_z.dfsu"
-    ds = mikeio.read("tests/testdata/oresund_sigma_z.dfsu", time=[0, 1])
-    ds.to_dfs(outfilename)
-
-    ds2 = mikeio.read(outfilename)
-    assert ds2.n_timesteps == 2
-
-
-def test_dataset_write_dfsu3d_max(tmp_path):
-
-    outfilename = tmp_path / "oresund_sigma_z.dfsu"
-    ds = mikeio.read("tests/testdata/oresund_sigma_z.dfsu")
-    assert ds._zn is not None
-    ds_max = ds.max("time")
-    assert ds_max._zn is not None
-    ds_max.to_dfs(outfilename)
-
-    ds2 = mikeio.read(outfilename)
-    assert ds2.n_timesteps == 1
-    assert ds2.geometry.is_layered
 
 
 def test_dataset_interp():
